refactor: extracted UI into another class
